requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
environs==11.0.0
marshmallow==3.21.1
packaging==24.0
pillow==10.2.0
python-dotenv==1.0.1
pytz==2023.3.post1
sqlparse==0.4.4
